Move util methods to HBaseRestUtils
